Add node-opencv patches
The files patches/VideoCapture.{cc,h}.patch can be used to consume
older frames in the video capture buffers, essentially turning the capture
stack into a Last-In-First-Out queue (vs. the default of First-In-First-Out)

It does the above by spinning off a thread which consumes frames as they
become available. Only one frame reference is persisted, which is then
provided to callers when requested.

Signed-off-by: James Ketrenos <james.p.ketrenos@intel.com>